+++ app/controllers/services/configurations_controller.rb
@@ -4,16 +4,19 @@ class Services::ConfigurationsController < Services::ApplicationController
   before_action :ensure_in_session!
 
   def show
-    if ProjectItem::Wizard::OfferSelectionStep.new(session[session_key]).valid?
-      @step = step_class.new(session[session_key])
+    if prev_step.valid?
+      @step = step(saved_state)
+
+      unless @step.visible?
+        redirect_to url_for([@service, next_step_key])
+      end
     else
-      redirect_to service_offers_path(@service),
-                  alert: "Please select one of the service offer"
+      redirect_to url_for([@service, pref_step_key]), alert: prev_step.error
     end
   end
 
   def update
-    @step = step_class.new(configuration_params)
+    @step = step(configuration_params)
 
     if @step.request_voucher
       @step.voucher_id = ""
@@ -21,7 +24,7 @@ class Services::ConfigurationsController < Services::ApplicationController
 
     if @step.valid?
       save_in_session(@step)
-      redirect_to service_summary_path(@service)
+      redirect_to url_for([@service, next_step_key])
     else
       flash.now[:alert] = @step.error
       render :show
@@ -30,13 +33,13 @@ class Services::ConfigurationsController < Services::ApplicationController
 
   private
     def configuration_params
-      template = ProjectItem.new(session[session_key])
-      session[session_key].
-          merge(permitted_attributes(template)).
-          merge(status: :created)
+      template = ProjectItem.new(saved_state)
+      saved_state
+        .merge(permitted_attributes(template))
+        .merge(status: :created)
     end
 
-    def step_class
-      ProjectItem::Wizard::ConfigurationStep
+    def step_key
+      :configuration
     end
 end