+++ app/models/project_item/wizard.rb
@@ -88,6 +88,8 @@ class ProjectItem::Wizard
       include ProjectItem::ProjectValidation
       delegate :properties?, to: :project_item
 
+      attr_accessor :additional_comment
+
       def error
         "Please select the project where service will be added"
       end