+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Catalog can be added to project (@martaswiatkowska)
 - Project_new_message mails content (@martaswiatkowska)
 - Information step to ordering path (@martaswiatkowska)
+- Use service type as default order type while seeding db (@mkasztelnik)
 
 ### Changed
 - Internal server error message (@goreck888)