+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Information step to ordering path (@martaswiatkowska)
 - Use service type as default order type while seeding db (@mkasztelnik)
 - Offer type to the project and offer selection views (@goreck888, @jarekzet)
+- Add help static page with example help accordion (@mkasztelnik)
 
 ### Changed
 - Internal server error message (@goreck888)