+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Project selection was moved to last service order step (@mkasztelnik)
 - Order configuration step is shown when offer is voucherable or has properties (@mkasztelnik)
 - Improve order wizard naming (@mkasztelnik)
+- Highlights in search from bold to mark (@goreck888)
 
 ### Deprecated